requirements.txt
fastapi==0.104.1
uvicorn==0.35.0
scikit-learn==1.5.1
joblib==1.4.2
pydantic==2.11.7
openai==2.14.0
python-dotenv==1.2.1
numpy==1.26.4
pandas==2.3.2
httpx==0.28.1
boto3
